Staff elements only visible if staff check passes
