+++ src/router/staffRoutes.ts
@@ -2,12 +2,12 @@ const staffRoutes = [
   {
     path: "/staff_area",
     name: "StaffArea",
-    component: () => import("@/views/StaffAreaView.vue"),
+    component: () => import("@/views/Staff/StaffAreaView.vue"),
   },
   {
     path: "/staff_area/applications_review",
     name: "ApplicationsReview",
-    component: () => import("@/views/ApplicationsReviewView.vue"),
+    component: () => import("@/views/Staff/ApplicationsReviewView.vue"),
   },
 ];
