--- src/views/ApplicationsReviewView.vue
@@ -1,5 +1,0 @@
-<template>
-  <p>Applications Review</p>
-</template>
-
-<script setup lang="ts"></script>
