+++ src/views/Staff/ApplicationsReviewView.vue
@@ -1,0 +1,16 @@
+<template>
+  <div v-if="isStaff">
+    <p>Applications Review</p>
+  </div>
+</template>
+
+<script setup lang="ts">
+import ApiClient from "@/ts/ApiClient";
+import { ref } from "vue";
+
+const isStaff = ref(false);
+
+ApiClient.loadPlayerInfo().then((player) => {
+  isStaff.value = player.isStaff;
+});
+</script>
