--- src/views/StaffAreaView.vue
@@ -1,5 +1,0 @@
-<template>
-  <p>Staff area</p>
-</template>
-
-<script setup lang="ts"></script>
